+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Upd Bask v0.5 (zstd compression, better support for wifi...)
 * Config for systemd-resolved, enable DNS over TLS with Quad9 (9.9.9.9)
 * Add configs for systemd-network with DHCP for wifi and ethernet.
 * Correct permissions (/home/[user] and /etc/portage)